// src/app/models/employee.model.ts
export class Employee {
    id?: number;
    name?: string;
    gender?:string;
    email?: string;
    phoneNumber?: string;
    contactPreference?: string;
    dateOfBirth?: string;
    department?: string;
    isActive?: boolean;
    photoPath?: string;
    password?:string;
    confirmpassword?:string;
}